Move Heartbeat Monitor into its own module
